refactor: split Wallet dialogs
