+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Wallet/DeleteConfirm.tsx
@@ -1,0 +1,45 @@
+import { Button } from '@material-ui/core'
+import { CreditCard as CreditCardIcon } from 'react-feather'
+import { WalletRPC } from '../../../../plugins/Wallet/messages'
+import { useI18N } from '../../../../utils/i18n-next-ui'
+import { DebounceButton } from '../../DashboardComponents/ActionButton'
+import SpacedButtonGroup from '../../DashboardComponents/SpacedButtonGroup'
+import { DashboardDialogCore, DashboardDialogWrapper, useSnackbarCallback, WrappedDialogProps } from '../Base'
+import type { WalletProps } from './types'
+
+export function DashboardWalletDeleteConfirmDialog(props: WrappedDialogProps<WalletProps>) {
+    const { t } = useI18N()
+    const { wallet } = props.ComponentProps!
+    const onConfirm = useSnackbarCallback(
+        async () => {
+            return WalletRPC.removeWallet(wallet.address)
+        },
+        [wallet.address],
+        props.onClose,
+    )
+    return (
+        <DashboardDialogCore fullScreen={false} {...props}>
+            <DashboardDialogWrapper
+                size="small"
+                icon={<CreditCardIcon />}
+                iconColor="#F4637D"
+                primary={t('delete_wallet')}
+                secondary={t('delete_wallet_hint')}
+                footer={
+                    <SpacedButtonGroup>
+                        <DebounceButton
+                            variant="contained"
+                            color="danger"
+                            onClick={onConfirm}
+                            data-testid="confirm_button">
+                            {t('confirm')}
+                        </DebounceButton>
+                        <Button variant="outlined" color="inherit" onClick={props.onClose}>
+                            {t('cancel')}
+                        </Button>
+                    </SpacedButtonGroup>
+                }
+            />
+        </DashboardDialogCore>
+    )
+}
