+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Wallet/ERC20PredefinedTokenSelector.tsx
@@ -1,0 +1,65 @@
+import { Box, createStyles, makeStyles, TextField } from '@material-ui/core'
+import { useState } from 'react'
+import { useI18N } from '../../../../utils/i18n-next-ui'
+import type { ERC20TokenDetailed } from '../../../../web3/types'
+import { EthereumTokenType } from '../../../../web3/types'
+import { FixedTokenList } from '../../DashboardComponents/FixedTokenList'
+
+const useERC20PredefinedTokenSelectorStyles = makeStyles((theme) =>
+    createStyles({
+        list: {
+            scrollbarWidth: 'none',
+            '&::-webkit-scrollbar': {
+                display: 'none',
+            },
+        },
+        search: {
+            marginBottom: theme.spacing(1),
+        },
+        placeholder: {
+            textAlign: 'center',
+            paddingTop: theme.spacing(10.5),
+            paddingBottom: theme.spacing(10.5),
+        },
+    }),
+)
+
+interface ERC20PredefinedTokenSelectorProps {
+    excludeTokens?: string[]
+    onTokenChange?: (next: ERC20TokenDetailed | null) => void
+}
+
+export function ERC20PredefinedTokenSelector(props: ERC20PredefinedTokenSelectorProps) {
+    const { t } = useI18N()
+    const classes = useERC20PredefinedTokenSelectorStyles()
+
+    const { onTokenChange, excludeTokens = [] } = props
+    const [keyword, setKeyword] = useState('')
+
+    return (
+        <Box
+            sx={{
+                textAlign: 'left',
+            }}>
+            <TextField
+                className={classes.search}
+                label={t('add_token_search_hint')}
+                autoFocus
+                value={keyword}
+                onChange={(e) => setKeyword(e.target.value)}
+                variant="outlined"
+            />
+            <FixedTokenList
+                classes={{ list: classes.list, placeholder: classes.placeholder }}
+                keyword={keyword}
+                blacklist={excludeTokens}
+                onSubmit={(token) => token.type === EthereumTokenType.ERC20 && onTokenChange?.(token)}
+                FixedSizeListProps={{
+                    height: 192,
+                    itemSize: 52,
+                    overscanCount: 2,
+                }}
+            />
+        </Box>
+    )
+}
