+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Wallet/Rename.tsx
@@ -1,0 +1,60 @@
+import { Button, TextField } from '@material-ui/core'
+import { useState } from 'react'
+import { WalletRPC } from '../../../../plugins/Wallet/messages'
+import { WALLET_OR_PERSONA_NAME_MAX_LEN } from '../../../../utils/constants'
+import { useI18N } from '../../../../utils/i18n-next-ui'
+import { checkInputLengthExceed } from '../../../../utils/utils'
+import { DebounceButton } from '../../DashboardComponents/ActionButton'
+import SpacedButtonGroup from '../../DashboardComponents/SpacedButtonGroup'
+import { DashboardDialogCore, DashboardDialogWrapper, useSnackbarCallback, WrappedDialogProps } from '../Base'
+import type { WalletProps } from './types'
+
+export function DashboardWalletRenameDialog(props: WrappedDialogProps<WalletProps>) {
+    const { t } = useI18N()
+    const { wallet } = props.ComponentProps!
+    const [name, setName] = useState(wallet.name ?? '')
+    const renameWallet = useSnackbarCallback(
+        () => WalletRPC.renameWallet(wallet.address, name),
+        [wallet.address],
+        props.onClose,
+    )
+    return (
+        <DashboardDialogCore fullScreen={false} {...props}>
+            <DashboardDialogWrapper
+                size="small"
+                primary={t('wallet_rename')}
+                content={
+                    <TextField
+                        helperText={
+                            checkInputLengthExceed(name)
+                                ? t('input_length_exceed_prompt', {
+                                      name: t('wallet_name').toLowerCase(),
+                                      length: WALLET_OR_PERSONA_NAME_MAX_LEN,
+                                  })
+                                : undefined
+                        }
+                        required
+                        autoFocus
+                        label={t('wallet_name')}
+                        value={name}
+                        onChange={(e) => setName(e.target.value)}
+                        inputProps={{ onKeyPress: (e) => e.key === 'Enter' && renameWallet() }}
+                    />
+                }
+                footer={
+                    <SpacedButtonGroup>
+                        <DebounceButton
+                            variant="contained"
+                            onClick={renameWallet}
+                            disabled={name.length === 0 || checkInputLengthExceed(name)}>
+                            {t('ok')}
+                        </DebounceButton>
+                        <Button variant="outlined" color="inherit" onClick={props.onClose}>
+                            {t('cancel')}
+                        </Button>
+                    </SpacedButtonGroup>
+                }
+            />
+        </DashboardDialogCore>
+    )
+}
