+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Wallet/index.tsx
@@ -1,0 +1,14 @@
+import './types'
+export { DashboardWalletAddERC20TokenDialog } from './AddERC20Token'
+export { DashboardWalletAddERC721TokenDialog } from './AddERC721Token'
+export { DashboardWalletBackupDialog } from './Backup'
+export { DashboardWalletDeleteConfirmDialog } from './DeleteConfirm'
+export { ERC20PredefinedTokenSelector } from './ERC20PredefinedTokenSelector'
+export { DashboardWalletErrorDialog } from './Error'
+export { DashboardWalletHideTokenConfirmDialog } from './HideTokenConfirm'
+export { DashboardWalletHistoryDialog } from './History'
+export { DashboardWalletImportDialog } from './Import'
+export { DashboardWalletRedPacketDetailDialog } from './RedPacketDetail'
+export { DashboardWalletRenameDialog } from './Rename'
+export { DashboardWalletTransferDialogFT } from './TransferFT'
+export { DashboardWalletTransferDialogNFT } from './TransferNFT'
