+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Wallet/types.ts
@@ -1,0 +1,5 @@
+import type { WalletRecord } from '../../../../plugins/Wallet/database/types'
+
+export interface WalletProps {
+    wallet: WalletRecord
+}
